% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildlnL.R
\name{buildlnL}
\alias{buildlnL}
\title{Function that builds lnL function based on dataMPN}
\usage{
buildlnL(d)
}
\arguments{
\item{d}{Data frame that is passed to the log likelihood function}
}
\description{
Builds a Negative log likelihood (lnL) function
}
