% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getALL.R
\name{getALL}
\alias{getALL}
\title{getALL function}
\usage{
getALL()
}
\arguments{
\item{theDF}{data.frame containing data according ot the template}
}
\description{
Function that allows the simultaneous MPN calculation according to the template
}
