% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.default.R
\name{getK.default}
\alias{getK.default}
\title{getK.default function}
\usage{
\method{getK}{default}(x, ...)
}
\arguments{
\item{x}{checking the class}

\item{...}{following objects}
}
\value{
an error message
}
\description{
Default function of getK, error message if the class does not corresponds
}
