% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.getK.R
\name{getK.getK}
\alias{getK.getK}
\title{This is the getK S3 function for class "getK" objects
 @export}
\usage{
getK.getK(..., timeVECT)
}
\description{
This is the getK S3 function for class "getK" objects
 @export
}
