% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.getMPN.R
\name{getK.getMPN}
\alias{getK.getMPN}
\title{This is the getK S3 function for class "getMPN" objects
 @export}
\usage{
getK.getMPN(..., timeVECT)
}
\description{
This is the getK S3 function for class "getMPN" objects
 @export
}
