% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.MPN.R
\name{getK.MPN}
\alias{getK.MPN}
\title{This is the getK S3 function for class "getMPN" objects}
\usage{
\method{getK}{MPN}(..., timeVECT)
}
\description{
This is the getK S3 function for class "getMPN" objects
}
