% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPN.R
\name{getMPN}
\alias{getMPN}
\title{Most Probable Number (MPN) function}
\usage{
getMPN(x, n, v, d = 10)
}
\arguments{
\item{x}{A numeric vector containing values of the positive wells.}

\item{n}{Number of MPN replicates}

\item{v}{Starting volume in ml or vector of volume}

\item{d}{Dilution factor, 10 as default}
}
\value{
data frame containing the MPN cytopathic unit per ml (MPNCU.ml), Upper and Lower 95% confidence interval (U_95CI and L_95CI), and standard error (SE)
}
\description{
Calculates the MPN according to a data.frame (or list).
}
