% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.getK.R
\name{plot.getK}
\alias{plot.getK}
\title{plotting function of getMPN}
\usage{
\method{plot}{getK}(...)
}
\arguments{
\item{object}{One of more getK objects}
}
\value{
plot
}
\description{
drawns a plot from getK or getMPN objects
}
