% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.getMPN.R
\name{plot.getMPN}
\alias{plot.getMPN}
\title{plotting function of getMPN}
\usage{
plot.getMPN(...)
}
\arguments{
\item{object}{One of more getMPN objects}
}
\value{
plot
}
\description{
drawns a plot from getMPN objects
}
