% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.K.R
\name{plot.K}
\alias{plot.K}
\title{Most Probable Number (MPN) function}
\usage{
plot.K(Object, type = "ln")
}
\arguments{
\item{Object}{K object}

\item{type}{y axis transformation}
}
\value{
a plot
}
\description{
Calculates the MPN
}
