% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPN.R
\name{plot.MPN}
\alias{plot.MPN}
\title{plot.MPN function}
\usage{
plot.MPN(..., plot = TRUE)
}
\arguments{
\item{...}{MPN objects}

\item{plot}{if =TRUE returns a dataframe, =FALSE returns a plot}
}
\value{
a dataframe or a plot
}
\description{
Plots the MPN calculated with the getMPN function
}
