% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPN.R
\name{plotMPN}
\alias{plotMPN}
\title{plotting function of getMPN}
\usage{
plotMPN(..., plot = TRUE)
}
\arguments{
\item{...}{One of more getMPN objects}
}
\value{
plot
}
\description{
drawns a plot from getMPN objects
}
