% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.K.R
\name{print.K}
\alias{print.K}
\title{print.K function}
\usage{
\method{print}{K}(x)
}
\arguments{
\item{x}{a K object output}
}
\value{
a dataframe of the results
}
\description{
S3 function to print only the results without the raw.data
}
