% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MPN.R
\name{print.MPN}
\alias{print.MPN}
\title{print.MPN function}
\usage{
\method{print}{MPN}(x)
}
\arguments{
\item{x}{a MPN object output}
}
\value{
a dataframe of the results
}
\description{
S3 function to print only the results without the raw.data
}
