% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.R
\name{getK}
\alias{getK}
\title{GetK function}
\usage{
getK(..., timeVECT, typeOF = FALSE)
}
\arguments{
\item{...}{One of more getMPN OF getk objects}

\item{timeVECT}{vector containing the time/dose points (necessary if getMPN as objects)}

\item{typeOF}{FALSE for getMPN combination and TRUE for getK combination.}
}
\description{
GetK function
}
