% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.R
\name{getK}
\alias{getK}
\title{getK function}
\usage{
getK(...)
}
\arguments{
\item{...}{One of more MPN or K objects}

\item{timeVECT}{vector containing the time/dose points (necessary if MPN objects)}
}
\description{
Calculate the inactivation constant k by combining MPN values and its uncertainties
Calculate the inactivation constant k by combining k values of different biological replicates
}
