% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPN.R
\name{plot.MPN}
\alias{plot.MPN}
\title{Most Probable Number (MPN) function}
\usage{
\method{plot}{MPN}(..., plot = TRUE)
}
\arguments{
\item{...}{MPN objects}

\item{plot}{if =TRUE returns a dataframe, =FALSE returns a plot}
}
\value{
a dataframe or a plot
}
\description{
Calculates the MPN
}
