% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotK.R
\name{plotK}
\alias{plotK}
\title{plotting function of getMPN}
\usage{
plotK(Object, type = "ln")
}
\arguments{
\item{Object}{A getK object}

\item{type}{"ln" to return a plot with natural logarithm plot, "log10" to return a log10 plot}
}
\value{
"ln" or "log10" plot
}
\description{
drawns a plot from getK objects
}
