% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPN.R
\name{plotMPN}
\alias{plotMPN}
\title{plotting function of getMPN}
\usage{
plotMPN(...)
}
\arguments{
\item{...}{One of more getMPN objects}

\item{yscale}{select the scale log10 or log}
}
\value{
plot
}
\description{
drawns a plot from getMPN objects
}
